Isolate the cams_solar_rad functions into a separate file
