+++ cads_adaptors/adaptors/cds.py
@@ -131,8 +131,8 @@ class AbstractCdsAdaptor(AbstractAdaptor):
             )
         # Avoid the cache by adding a random key-value pair to the request (if cache avoidance is on)
         if self.config.get("avoid_cache", False):
-            random_key = str(randint(0,1e30))
-            request['_in_adaptor_no_cache'] = random_key
+            random_key = str(randint(0, 1e30))
+            request["_in_adaptor_no_cache"] = random_key
         return request
 
     def get_licences(self, request: Request) -> list[tuple[str, int]]:
